video_processor.py

import json
from typing import List, Dict
import cv2  # OpenCV for video processing
from PIL import Image  # Pillow for image handling
import matplotlib.pyplot as plt
import joblib  # For saving the model

# Define data classes
class VideoObject:
    def __init__(self, path: str):
        self.path = path
        self.frames = []

class Frame:
    def __init__(self, index: int, image):
        self.index = index
        self.image = image

class AnalysisResults:
    def __init__(self, features: Dict):
        self.features = features

# Define the VideoProcessor class
class VideoProcessor:
    def load_video(self, video_path: str) -> VideoObject:
        video = VideoObject(video_path)
        return video

    def extract_frames(self, video_object: VideoObject, method: str, parameters: Dict) -> List[Frame]:
        cap = cv2.VideoCapture(video_object.path)
        frames = []
        index = 0

        if method == 'fixed_interval':
            interval = parameters.get('interval', 30)
            while cap.isOpened():
                ret, frame = cap.read()
                if not ret:
                    break
                if index % interval == 0:
                    frames.append(Frame(index, frame))
                index += 1

        cap.release()
        video_object.frames = frames
        return frames

    def analyze_features(self, frames: List[Frame]) -> AnalysisResults:
        features = {'total_frames': len(frames)}
        for frame in frames:
            avg_color = cv2.mean(frame.image)[:3]
            features[f'frame_{frame.index}_avg_color'] = avg_color
        return AnalysisResults(features)

    def save_results(self, results: AnalysisResults, path: str) -> None:
        with open(path, 'w') as f:
            json.dump(results.features, f, indent=4)

# Example usage
if __name__ == "__main__":
    processor = VideoProcessor()
    video = processor.load_video('video/video1.mp4')
    frames = processor.extract_frames(video, 'fixed_interval', {'interval': 30})
    results = processor.analyze_features(frames)
    processor.save_results(results, 'results/results.json')
    joblib.dump(processor, 'results/video_processor.pkl')
    