requirements.txt
joblib
opencv-python
pillow
matplotlib
